Implement separate server_socket and socket classes

+++ src/addrinfo.hpp
@@ -4,13 +4,14 @@
 #include <uv.h>
 #include <memory>
 #include <string>
+#include <cstdint>
 
 namespace fiberio {
 
 using addrinfo_ptr =
     std::unique_ptr<struct addrinfo, decltype(&uv_freeaddrinfo)>;
 
-addrinfo_ptr getaddrinfo(const std::string& host, int port);
+addrinfo_ptr getaddrinfo(const std::string& host, uint16_t port);
 
 addrinfo_ptr getaddrinfo(const std::string& node, const std::string& service);
